changing modalEdit funct to new file
